backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
hypothesis>=6.92.1
httpx>=0.25.2
pillow>=10.0.0
